[Fix] Bundle scattered data for coherence
